+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/AutoCrafterListener.java
@@ -2,7 +2,6 @@ package io.github.thebusybiscuit.slimefun4.implementation.listeners;
 
 import java.util.Optional;
 
-import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.bukkit.GameRule;
@@ -20,7 +19,9 @@ import io.github.thebusybiscuit.slimefun4.api.events.PlayerRightClickEvent;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.items.autocrafters.AbstractAutoCrafter;
 import io.github.thebusybiscuit.slimefun4.implementation.items.autocrafters.EnhancedAutoCrafter;
+import io.github.thebusybiscuit.slimefun4.implementation.items.autocrafters.VanillaAutoCrafter;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.gadgets.Multimeter;
+
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 /**
@@ -30,6 +31,9 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
  * This {@link Listener} resolves that issue.
  *
  * @author TheBusyBiscuit
+ * @author LilBC
+ * 
+ * @see VanillaAutoCrafter
  * @see EnhancedAutoCrafter
  */
 public class AutoCrafterListener implements Listener {
@@ -37,15 +41,6 @@ public class AutoCrafterListener implements Listener {
     @ParametersAreNonnullByDefault
     public AutoCrafterListener(SlimefunPlugin plugin) {
         plugin.getServer().getPluginManager().registerEvents(this, plugin);
-    }
-
-    private boolean hasUnlockedRecipe(@Nonnull Player p, @Nonnull ItemStack item) {
-        for (Recipe recipe : SlimefunPlugin.getMinecraftRecipeService().getRecipesFor(item)) {
-            if (recipe instanceof Keyed && !p.hasDiscoveredRecipe(((Keyed) recipe).getKey())) {
-                return false;
-            }
-        }
-        return true;
     }
 
     @EventHandler
@@ -74,10 +69,15 @@ public class AutoCrafterListener implements Listener {
                 // Prevent blocks from being placed, food from being eaten, etc...
                 e.cancel();
 
-                // Check if the recipe of the item is disabled.
-                if (e.getPlayer().getWorld().getGameRuleValue(GameRule.DO_LIMITED_CRAFTING) && !hasUnlockedRecipe(e.getPlayer(), e.getItem())) {
-                    SlimefunPlugin.getLocalization().sendMessage(e.getPlayer(), "messages.auto-crafting.recipe-unavailable");
-                    return;
+                // Check for the "doLimitedCrafting" gamerule when using a Vanilla Auto-Crafter
+                if (block instanceof VanillaAutoCrafter) {
+                    boolean doLimitedCrafting = e.getPlayer().getWorld().getGameRuleValue(GameRule.DO_LIMITED_CRAFTING);
+
+                    // Check if the recipe of the item is disabled.
+                    if (doLimitedCrafting && !hasUnlockedRecipe(e.getPlayer(), e.getItem())) {
+                        SlimefunPlugin.getLocalization().sendMessage(e.getPlayer(), "messages.auto-crafting.recipe-unavailable");
+                        return;
+                    }
                 }
 
                 // Fixes 2896 - Forward the interaction before items get handled.
@@ -90,5 +90,16 @@ public class AutoCrafterListener implements Listener {
                 }
             }
         }
+    }
+
+    @ParametersAreNonnullByDefault
+    private boolean hasUnlockedRecipe(Player p, ItemStack item) {
+        for (Recipe recipe : SlimefunPlugin.getMinecraftRecipeService().getRecipesFor(item)) {
+            if (recipe instanceof Keyed && !p.hasDiscoveredRecipe(((Keyed) recipe).getKey())) {
+                return false;
+            }
+        }
+
+        return true;
     }
 }